Add Coggan style world ranking to crit power info.
I wrote a horrible equation to try and map the Andy Coggan power
profile spreadsheet values into a dynamic range.  With this you can
estimate where your crit power values fall in comparison to the best
capabilities of the worlds best or the best performance of someone
that is untrained (recreational).  More info:
  https://www.cyclinganalytics.com/blog/2012/06/watts-kg-on-the-power-curve