More loading race-cond fixes, enhancements.
The code to avoid racing with in-flight streamRequests was racy given
that the deferred object being used is replaced after a succussful call.
Instead we need a more in-depth test for in-flight conditions.  This
adds such a test so we load regardless of what is going on with the
network.

Templatize the critpower panel.

Various jshint fixes.

Bump browser requirement to chrome 51 which added stable "for of"
support.