Working to handle moreinfo dialogs for all sources.
We can't depend on the moreinfo dialog being started by just one source.
So we need to start it with just timestamps and build everything from
scratch.

Also working to support these dialogs without power data or other stream
sources I was previously assuming were required.  For example we could
show the dialog for an activity with just HR (and no distance if indoors)