Fix bug with rides that don't have power data / Analysis menu fixes
The analysis menu item needs to be present early on so it can receive
class updates (e.g. .selected).  So I add the element as a stub and hide
it if it's not there.  Then later in the analysis page it looks for the
stub and fixes it up as needed.   Now reloading on the analysis page has
the correct selection class set.