Converging on unified RPC mech.
This is sort of nice actually.  I'm able to drop the requirement for
external connectability.   Instead I just use standard
window.postMessage to do RPC to the ext content script.  If the RPC hook
is "backgroundOnly" then I forward the request to the background page.

This is only required for a few things.  For GA because I need some tab
info and for cases where I want to enforce atomic actions like patching
storaga values (requires locks).