Fix FTP loading for one rare condition / fix some races.
If you WERE paying and set your FTP then cancelled your FTP would still
be in the system and can be found in the non-full activity.  I was using
the "full" activity in nearly all cases.  I noticed a couple other
issues with the "full" activity, so I'm demoting it to specific use
cases where I know we might need it (like getting the activity time).

Related to #7