Guard slideMenu is null error.
I can't figure out why but sometimes .slide-menu is not in the DOM even
when nav.sidenav is.  This makes no sense to me because they come in as
the same server side generated HTML template in every case I've
examined.

So I'm adding a dumb guard since I can't figure it out.