Fix fetchstreams hang bug with empty required.
Our fetchStreams routine attempts to debounce pending requests.  However
we started the pending state even when the "required" streams set was
empty.  In this case the pending state would never be cleared because
the StreamRequest code would know that no work was needed.  This would
cause fetchStreams to hang indefinetly.

Shows up on certain analysis pages like "Heartrate" if you reload the
site on this page.