Handle popup and options as different ux.
Since popup actions are basically being hidden by Chrome now, I need to
provide a better options experience and the best way of doing this is to
make the options pages full tabs.   So the pages need to be tuned for
this, style wise and for not attempting to reload the active tab, since
it won't exist in the options page context.