More Caching and more aggressive preloader loading.
Despite the preloader being loaded as early as browser extensions allow
it is subject to network latency that permits inline source scripts to
be loaded before preloader.js is evaluated and executed.  This is even
with script.async = false being set on the script element.  So the only
migigation I've found that works 100% of the time is to load the
preloader and base code as an anonymous script.  To do this I must load
these files into the extension content_script context and then copy
their function contents (e.g. Function.toString()) into an inline script
tag.  This blocks the site page from loading until the preloader is
fully evaluated.

Caching is now implemented for segment leaderboards and segment effort
details.  This makes strava tremendously faster when bouncing around
within one activity or after a quick reload of the site.