Bravely moving forward with the XXX DRY refactor
This is pretty close to FC, but lots of testing and tweaks for range
based info dialogs like pace;  As well as Speed vs Pace headings and
ordering for things.  Also the info dialog itself is a complete overhaul
work in progress but I've gotten positive feedback about the direction
so I'm committed now.