FC on new info dialogs.
This is feeling really close to me now.  I'm liking it finally at least.

Segment overlap is now working well and the expander is stateful between
page reloads (and within a page).

Lots of style fixes and handling of our width changes.

Proper theme support for gradients and other colors.

Locale for strings (not translated)