Fix TCX export regression.
Lap StartTime was broken, using the last time stamp of the lap instead
of the first because I failed to recognize the attribute was being set
in the inner track point loop.  Woops!

Also the lapEfforts collection must be loaded at export time since it's
a network resource.  This led to race conditions with lap data being
present or not.

Fixes #15