Mostly working proxy system.
The jist of this refactor is to remove a lot of boilerplate when adding
proxy functions (previosuly called RPC).  And the protocol has been
expanded a bit to support eventing in the next iteration.  This will
allow me to setup clean eventing between the site context and the other
extension contexts such as the background page.  Ie. Keeping tabs on
history sync status and what not, without need for polling.