More reliable current user detection (but slower).
The localStorage(ajs_user_id) is sometimes not present or maybe even
deprecated, so I'm not just letting the analysis page find the current
user based on the pageView.currentAthlete.  This is slower because we
can't detect the cur user until you visit an activity page.  Then you'll
have to refresh the page to get it sent out to the ext and bg contexts.

I can probably remove that last hurdle with an event from the analysis
page but for now this will do since we don't depend on it too heavily.