Remove use of localStorage for patron detection / en/dis bg support
With our updated user ID detection there is no reason to try and use
localstorage anymore.  Originally it saved us from a first use detection
issue, but it's unavoidable now.  So just simplify the heck out of it
with sauce.storage.

Removed the unjustifiably small patron.js file.

Added support for enabling and disabling the background page.  Now that
it does actual background work, we need to instruct it to stop or start
based on the user's enabled/disabled pref.  This could be a VERY useful
fuse in the future if we cause some sort of problem for the user doing
syncs.