Design issue!
The training load processor has to modify activities outside the batch
it's given to process.  And the models being modified are being
processed in other batches.  Presently it's causing data corruption
issues.  Need a better way to instruct the processor that some manfiest
jobs need to run with the totality of activities and or have a way to
interact with the sync job.  Ie. instruct it to refresh some activities
or to commit changes on behalf of the processor.  A la an intent system.