Checkpoint: Working on yet-another refacotor of local proc.
Need a way to genericly support offloading activity processing without
exiting the sync-job.  So now processors can return a queue.Queue object
which is used to indicate that the procesing will happen out of band
and the sync engine is free to ingest more work.  It won't howerver,
proceed with work related to those activiites.  So essentially it will
allow for a processor to do 2 things:

    1. Wait for more activiites to have been fetched from the network.
    2. Wait for more activities to have been promoted from dependant
       or otherwise busy processors working.

By being able to wait out of band, the processor can be as creative (or
simple) as it wants with respect to buffering, debouncing, etc.