Move offload processor to class based design.
We need to track errors and there are other more complicated life cycle
things that I'd prefer to keep out of the business logic, so now if you
want an offload processor you need to subclass OffloadProcessor and fill
out the processor() function.  It's a one method interface, but you can
also get as fancy as you want.  If it blows up then the sync job handles
everything as you would expect.