Perf tuning.
Trying some alternative peak queries that get all the values in a time
range then sort and limit.  In every case I've seen it's MUCH faster.
However for multi athlete there could easily be a limit to the number of
athletes where this makes sense.  Knowing in advance may not be possible
in those cases without dong a query of unique athletes first.  Tough
call.