Monkey patch setTimeout for  wall-clock timer
setTimeout uses an internal timer that is suspended during OS sleep
states.  This has the effect of deferring our sleeps for arbitrary
amounts of time.  We always want our timers to be based on wall-clock
timers since they are calender based.  e.g. wake up at this time of the
day to re-sync.  In the ephemeral web world the spec behavior of
setTimeout is fine, but our background page is a long lived app.

This patch installed a setInterval loop that detects matured setTimeout
calls and invokes them as soon as possible.  For example.

  10:00 setInterval for effectively 20:00 (10 hours)
  11:00 OS enters sleep state
  21:00 OS leaves sleep state
  21:00:01 setInterval loop detects mature setInterval callback and runs
     it.

Without intervention the normal setInterval call would have seen 1
hour of elapsed timer, then slept for 10 hours.  Upon resume it would
see that it has a pending setInterval call that now has 9 full hours of
timer remaining before it will invoke it.  Even if the OS stayed awake
for this remaining time, the setInterval wouldn't be invoked until 06:00
the following DAY!

Fixes #27