Deploy amazing wow indexeddb perf fix for Safari
Safari has some issue with multiple read-only IDB transactions being
held open in webworkers.  They are serialized apparently, and adding a
sleep statement between the DB work and the CPU work forces the
transaction to close and our web workers can run in parrallel, as god
intended.

In my test of my data local processing is 300% faster with this on my
laptop.  I don't expect any perf change on the other browsers, but we
should check just to be sure.