CalendarRange and other byDay fixes / support for custom defaults.
Fixes for acts that start at exactly midnight (manual entry sometimes
does this if you don't spec a real time.  The dayRange iter is exclusive
on the upper bound so we need to ff to the start of the next day.

Support for entering a custom period and metric into the url and
actually having it get remembered as a default and even be selectable in
the time range select.  It's a bit of an easter egg, but it's a cool
feature and I can extend the UI to grant this power to laymen.

Fix for range changes that would foward the end time past the present
day.