Safer/stricter handling of ext storage updates.
Pretty glarring hole in storage.<type>.set(key, value) is that the
values are JSONified which turns values of `undefined` into a complete
noop.  Meaning that trying to clear a value by setting it to undefined
actually has no effect and the orignal value is persisted.  So we need
to specifically look for undefined values and explicitly call
storage.<type>.remove(...) with those keys.

The other change here is to force sauce.storage.update to type check
the incoming value as it MUST be an object.  I don't think I have code
abusing this, but in testing it became clear that it's handling that
case completely wrong, so I force an exception now, to make sure I don't
ever accidentially do this later (or in some corner case).