Fix/enhance ext context vs page context.
template code was just using a bad ref to browser.runtime.getURL.
Luckily we never use templates in ext ctx yet.  But some users have
other extensions that actualy produce a viable "browser" variable at
global scope.  So we need to make the test for "Are we in ext context"
more strict.