Make PageView.render() idempotent
DOM events would be removed from sync button on render due to the nature
of jQuery.html().   We need to detach or throwaway any elements managed
by jQuery during render of these type of views.  Opting to just recreate
the sync buttons in this case as they are very quick to create after the
template is loaded once.

Fixes #21