Add sync job timeout
I've not personally seen it, at least recently, but the threat of a
wedged sync job is high, so I'm adding a timeout to sync jobs that
will cancel and restart them if they are running for more than 90 mins.

Hopefully this only triggers for initial sync when we run into rate
limiter delays.  In these cases it's a no-op (other than the blip of
restarting).

Some minor refactoring to streamline event reporting and seperation
of concerns with SyncJob vs SyncManager.

Fixes #23