Fix dropped sync progress.
Now that sync progress includes a list of done activity stats we need to
send a progress event if it's populated.  The last round of processing
will likely include a final pass where the counts stay the same but the
done list isn't populated until the final iteration, so we need to send
one more progress event.