Micro and macro optimizations for templating and locales
Refine and offer speedups for sauce.locale functions.  Allow them to be
used synchronously if the caller is willing to eval the result and
look for a promise.

Optimizations for template rendering.  Pull all strings during
complilation instead of render.  Add support for "static helpers"
which are like helpers but use a new grammer and are parsed at compile
time.   The syntax is:

        {{=funcname funcargs=}}

The static helper, 'funcname' will be called with a string value
of 'funcargs' which it can process at compile time and the result is
saved for future renders.  Currently this supports icons, e.g..

        {{=icon circle-duotone=}}