Massive refactor checkpoint.
Using offloaded processor for more stuff and fixing some issues with it.
Still need to look into flush/shutdown mechanics.  They are not perfect.

Put activity updates into the syncjob class so they can use retryFetch
with a cancel event and get throttle handling.

Stagger sync job invocation when not initiated by the user.  So on a
version change for example we won't destroy the user's computer.