Don't use autoIncrement keys for TTLCache
Honestly I thought this was the the cause of a chrome perf issue for
idb, but it turns out idb on chrome is just horribly slow for a new
readwrite transaction (190ms if you can believe that).  However the
change is not a bad one.  The prior method required a lot more
extraneous code.