Use functional approach to range shifting (perf page)
This fixes various race conditions with fast navigation in the perf
page.  We pass a frozen snapshot of the range to the event listeners and
ensure that they always use the state from that object and don't reach
into the pageview unless it's completely safe.

Also streamlined the rendering of the data range itself, so it's painted
immediately.  Which incidentally fixes the ability to over shift (btn
becomes disabled immediately) and go into the future where not acts exist.

Fixes #65