Restore data / sync panel monitor or athlete record updates.
A user reported that they lost their weight history and it was likely
due to the sync panel enabled button being pressed on a fresh machine
after an import which would clobber the athlete record.

This should make it so athlete records will not be clobbered in any
context.  But also redraws various components of the sync panel as a
contextually relevant import is taking place.