Really naieve offload proc for peaks for testing.
This should be enhanced significantly but I wanted to test out the basic
concept of feeding workers and put finished results back into the sync
engine as fast as possible.  It's quite a bit faster on chrome and much
faster on firefox.   Making them almost equal for a full sync actually.