Yet more changes for RollingAverage - Prevent Pad bounds
When doing rolling power, trim off Pad boundries by default.  For Zwift
I'll disable this as we don't have complete data (it comes in and out)
but for Strava we need to make sure the first and last indexes map to
real data points since we don't store padded data (maybe we should?).