Revert recent rolling avg changes
Too much perf regression and it's too hard to correlate to zwifts
unpadded streams.  Going to bench this and try an approach that skips
pad values alltogether in favor of storing a meta data stream of gaps,
but it's going to take some major effort and testing.  Too much of
a distraction right now.