peaks-wkg -> peaks-finalizer
The w/kg/cleanup processor is now the peaks-finalizer processor.

It does proper cleanup of legacy (missing estimate flag) entries as well
as purging entries with obsolete periods and distances.

One thing of note, but not related to this commit.  The rankLevel is
only done in the finalizer which means this processor will be rewriting
all new entries created by the main peaks processor.  This is quite
wasteful so I might move rankLevel computation back to the main peaks
proc as well. Perhaps to the run-power proc too.  In which case, it
might revert into just a cleanup processor.