Use better strategy for day bucket filling in training load proc.
Instead of assuming all activities are in the present time frame use the
timestamp of the actual newest actvity seen for the dayRange iter.  If
the users clock is messed up or they have a wonky activity that claims
to be from the future, we'll avoid blowing up.

Also improved the internal error to a harmless report instead.

Minor tweak to dayRange to support ts and Date for both start and end.
Before it was only tollerant of a ts start.