Support for do not track.
It's a deprecated standard but it still is useful.  Firefox will even
set it when it's setup for strict browsing mode, so it's useful.

On top of that if we are consistently failing to send events when the
browser is claiming to be navigator.onLine assume that they have a
blocker enabled (other extension or builtin).  In that case treat it
the same as dnt, except it is forever.