Fix / enhance activity sync.
Include photos and more details non-self.  Calling the self update V2,
and not turning it on by default yet.  It's serial and so much much
slower than the concurrent original impl.  But it does have photos so
I'll have to work on it a bit and see if I can make it concurrent
safely.