More aggressive backoff for sync errors.
Some accounts like Phil G. return a ton of 500 errors when asking for
the intervals endpoint.  So many that our sync engine was trying to pull
the entire activity history on every sync because there were never more
than 2 redundant months in a row.

Increased backoffs in general.
Reduced max retries of an invidual fetch from 5 to 3.
Wire in cancel event to activity updating.