Fix mem overflow on big sync jobs / remove setTimeout monkey patch
When doing huge (20GB-ish) syncs we could run out of memory because the
peaks processor would try to consume all the pending work in every
batch.  Limit this to a reasonable chunk size and we're golden.

Removed monkey patch'd setTimeout and just plumb in the specific cases
were we need to aggressively sleep.