Restart interrupted sync jobs immediately.
Because service workers are going to die frequently we need to pick back
up any active syncs right away.  The old system was more forgiving and
would just wait until the next normal checkin deadline of 12 hours.  Now
we'll set lastSync to 0 before the sync job starts.  And only set it
back to a current date if the user manually stops or it finishes.