Revert timezone aware export / refactor time prompt
I'm worried about compat for timezone aware ISO8601 values in TCX and
GPX output for now.  It's not clear to me that this would indicate
anything to a client in the first place but I'm happy to be proven wrong
about that.

Some cleanup / refactor for datetime prompt to look for syncActivity.ts
first and then if all else fails ask the user for a datetime.

Removed the hack of using startDateLocal and then trying to guess
timezone.  Error prone and generates too many support calls.